refactor(backend): isolate Codex feature and catalog domains
