+++ src-tauri/src/codex_config.rs
@@ -15,6 +15,8 @@ use std::process::{Command, Stdio};
 use toml_edit::{DocumentMut, Item, TableLike};
 
 mod auth;
+mod catalog;
+mod features;
 mod storage;
 
 pub use auth::{
@@ -25,6 +27,30 @@ pub use auth::{
 #[cfg(test)]
 use auth::{
     codex_auth_has_credential_login_material, codex_live_auth_is_stale_third_party_residue,
+};
+
+#[cfg(test)]
+use catalog::{build_simplified_catalog_from_texts, read_limited_string, MAX_CODEX_CATALOG_BYTES};
+pub use catalog::{read_codex_model_catalog_simplified_from_live, CodexCatalogToolProfile};
+pub(crate) use catalog::{read_codex_model_catalog_text, resolve_fyagent_catalog_path};
+
+#[cfg(test)]
+pub use features::CODEX_WEBSOCKET_NON_GPT_MODEL_WARNING;
+use features::{
+    active_codex_provider_table, inspect_managed_image_header, set_managed_image_header,
+    ManagedHeaderInspection,
+};
+pub use features::{
+    analyze_codex_provider_features, codex_provider_save_warning_codes,
+    patch_codex_provider_features, prepare_codex_provider_features_for_save,
+    validate_codex_provider_features, CodexProviderFeatureIntent, CodexProviderFeaturePatchResult,
+    CodexProviderFeatureState, CODEX_IMAGE_EXTENSION_HEADER, CODEX_IMAGE_EXTENSION_VALUE,
+};
+#[cfg(test)]
+use features::{
+    codex_provider_config_text, set_provider_config_text, CodexImageExtensionState,
+    CODEX_FEATURE_INVALID_TOML, CODEX_FEATURE_INVALID_WEBSOCKET,
+    CODEX_WEBSOCKET_PROXY_MAY_BE_UNSUPPORTED_WARNING,
 };
 
 pub use storage::{
@@ -41,88 +67,6 @@ pub const FYAGENT_CODEX_MODEL_PROVIDER_ID: &str = "custom";
 pub const FYAGENT_CODEX_OFFICIAL_PROXY_PROVIDER_ID: &str = "fyagent-official";
 pub const FYAGENT_CODEX_MODEL_CATALOG_FILENAME: &str = "fyagent-model-catalog.json";
 const CODEX_PROXY_AUTH_PLACEHOLDER: &str = "PROXY_MANAGED";
-
-/// The provider-scoped request header owned by the FyAgent image-extension
-/// feature. Header names are semantically case-insensitive, while TOML keys
-/// are not, so every read and mutation must use the helper below instead of a
-/// direct TOML key lookup.
-pub const CODEX_IMAGE_EXTENSION_HEADER: &str = "x-openai-actor-authorization";
-pub const CODEX_IMAGE_EXTENSION_VALUE: &str = "local-image-extension";
-const CODEX_FEATURE_INVALID_TOML: &str = "CODEX_FEATURE_INVALID_TOML";
-const CODEX_FEATURE_INVALID_HEADER: &str = "CODEX_FEATURE_INVALID_HEADER";
-const CODEX_FEATURE_INVALID_WEBSOCKET: &str = "CODEX_FEATURE_INVALID_WEBSOCKET";
-pub const CODEX_WEBSOCKET_NON_GPT_MODEL_WARNING: &str = "CODEX_WEBSOCKET_NON_GPT_MODEL";
-pub const CODEX_WEBSOCKET_PROXY_MAY_BE_UNSUPPORTED_WARNING: &str =
-    "CODEX_WEBSOCKET_PROXY_MAY_BE_UNSUPPORTED";
-
-/// Serializable, non-sensitive diagnostic for the feature-draft IPC. Never
-/// include a header value or a complete TOML fragment here: provider headers
-/// may themselves contain credentials.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "camelCase")]
-pub struct CodexConfigDiagnostic {
-    pub code: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub field: Option<String>,
-}
-
-/// The effective image-extension state derived from the provider TOML plus
-/// the private migration marker. TOML remains the source of truth; the marker
-/// only distinguishes a historical missing header from an explicit disable.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(tag = "kind", rename_all = "camelCase")]
-pub enum CodexImageExtensionState {
-    On,
-    Off,
-    LegacyPendingOn,
-    Conflict { key: String },
-    Invalid { code: String },
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "camelCase")]
-pub struct CodexWebsocketFeatureState {
-    pub enabled: bool,
-    pub compatible: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub reason: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "camelCase")]
-pub struct CodexProviderFeatureState {
-    pub applicable: bool,
-    pub image_extension: CodexImageExtensionState,
-    pub websockets: CodexWebsocketFeatureState,
-    pub provider_table_found: bool,
-    pub diagnostics: Vec<CodexConfigDiagnostic>,
-}
-
-/// Intent sent by the draft-only feature patch command. `None` means leave
-/// that capability unchanged. The command operates on form text only and
-/// never writes a user file or database row.
-#[derive(Debug, Clone, Deserialize, Serialize, Default, PartialEq, Eq)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub struct CodexProviderFeatureIntent {
-    pub image_extension: Option<bool>,
-    pub websockets: Option<bool>,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, Eq)]
-#[serde(rename_all = "camelCase")]
-pub struct CodexProviderFeaturePatchResult {
-    pub toml_text: String,
-    pub state: CodexProviderFeatureState,
-    /// A proposed, private ProviderMeta update. The caller must persist this
-    /// only together with a successful provider save; cancelling a form must
-    /// therefore leave the migration marker absent.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub image_extension_configured: Option<bool>,
-    /// Form-local update for the private ownership marker of an official
-    /// capability-only provider table. `Some(false)` explicitly clears it.
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub codex_native_capabilities_generated_provider: Option<bool>,
-}
 
 #[cfg(target_os = "windows")]
 const CREATE_NO_WINDOW: u32 = 0x08000000;
@@ -216,46 +160,6 @@ fn codex_native_gateway_rejects_web_search(config_text: &str) -> bool {
 }
 const CODEX_MODEL_CATALOG_TEMPLATE_SLUG: &str = "gpt-5.5";
 
-/// Which Codex tool surface the generated model catalog should target.
-///
-/// - `ProxyChat`: fyagent's proxy takes over and converts Responses<->Chat,
-///   so the catalog keeps Codex's default tool set (incl. the freeform
-///   `apply_patch` custom tool, which the proxy rewrites to a function tool).
-/// - `NativeResponses`: Codex talks directly to a provider's native
-///   `/responses` endpoint (no proxy). Such gateways (e.g. Xiaomi MiMo,
-///   MiniMax) reject `type=="custom"` tools, so the catalog must suppress the
-///   freeform `apply_patch` and rely on `shell_type="shell_command"` for edits.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum CodexCatalogToolProfile {
-    ProxyChat,
-    NativeResponses,
-    /// Codex talks (through fyagent's proxy) to a native Anthropic Messages
-    /// gateway. Like `NativeResponses` it must suppress Codex's freeform custom
-    /// tools — the Responses→Anthropic transform keeps only `function` tools.
-    /// Additionally the Codex `web_search` hosted tool is unusable on this path
-    /// (the transform drops it), so it is always disabled — see
-    /// `prepare_codex_config_text_with_model_catalog`.
-    Anthropic,
-}
-
-impl CodexCatalogToolProfile {
-    /// Pick the catalog tool profile from a provider's `apiFormat` meta value.
-    ///
-    /// Prefer [`crate::proxy::providers::codex::resolve_codex_catalog_tool_profile`],
-    /// which also honors settings-level `apiFormat` and the TOML `wire_api` (matching
-    /// the proxy router). This string-only mapping is the fallback for non-Anthropic
-    /// cases.
-    pub fn from_api_format(api_format: Option<&str>) -> Self {
-        match api_format {
-            Some("anthropic") => CodexCatalogToolProfile::Anthropic,
-            // Native (direct) Responses gateways reject Codex's freeform custom
-            // tools (apply_patch, etc.); strip them via the NativeResponses profile.
-            Some("openai_responses") => CodexCatalogToolProfile::NativeResponses,
-            _ => CodexCatalogToolProfile::ProxyChat,
-        }
-    }
-}
-
 /// Reserved built-in provider IDs from OpenAI Codex's config/model-provider
 /// catalog. Keep in sync with Codex `RESERVED_MODEL_PROVIDER_IDS` and legacy
 /// removed provider aliases.
@@ -282,664 +186,6 @@ pub(crate) fn is_custom_codex_model_provider_id(id: &str) -> bool {
         && !CODEX_RESERVED_MODEL_PROVIDER_IDS
             .iter()
             .any(|reserved| reserved.eq_ignore_ascii_case(id))
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum ManagedHeaderInspection {
-    Missing,
-    Controlled { key: String },
-    Conflict { key: String },
-    Invalid,
-}
-
-fn codex_feature_error(code: &'static str, zh: &'static str, en: &'static str) -> AppError {
-    AppError::localized(code, zh, en)
-}
-
-fn codex_feature_diagnostic(code: &'static str, field: &'static str) -> CodexConfigDiagnostic {
-    CodexConfigDiagnostic {
-        code: code.to_owned(),
-        field: Some(field.to_owned()),
-    }
-}
-
-fn codex_provider_config_text(provider: &Provider) -> &str {
-    provider
-        .settings_config
-        .get("config")
-        .and_then(Value::as_str)
-        .unwrap_or_default()
-}
-
-fn active_codex_provider_table(doc: &DocumentMut) -> Option<(String, &dyn TableLike)> {
-    let provider_id = active_codex_model_provider_id(doc)?;
-    let table = doc
-        .get("model_providers")
-        .and_then(Item::as_table_like)?
-        .get(&provider_id)
-        .and_then(Item::as_table_like)?;
-    Some((provider_id, table))
-}
-
-fn active_codex_provider_table_mut<'a>(
-    doc: &'a mut DocumentMut,
-    provider_id: &str,
-) -> Option<&'a mut dyn TableLike> {
-    doc.get_mut("model_providers")
-        .and_then(Item::as_table_like_mut)?
-        .get_mut(provider_id)
-        .and_then(Item::as_table_like_mut)
-}
-
-fn is_fixed_official_codex_provider(provider: &Provider) -> bool {
-    provider
-        .id
-        .eq_ignore_ascii_case(crate::database::CODEX_OFFICIAL_PROVIDER_ID)
-        || provider
-            .category
-            .as_deref()
-            .is_some_and(|category| category.eq_ignore_ascii_case("official"))
-}
-
-fn generated_official_provider_marker(provider: &Provider) -> bool {
-    provider
-        .meta
-        .as_ref()
-        .and_then(|meta| meta.codex_native_capabilities_generated_provider)
-        == Some(true)
-}
-
-fn inspect_managed_image_header(provider_table: &dyn TableLike) -> ManagedHeaderInspection {
-    let Some(headers_item) = provider_table.get("http_headers") else {
-        return ManagedHeaderInspection::Missing;
-    };
-    let Some(headers) = headers_item.as_table_like() else {
-        return ManagedHeaderInspection::Invalid;
-    };
-
-    let mut matching_keys = Vec::new();
-    for (key, value) in headers.iter() {
-        // Codex http_headers must be a string map. Returning a generic
-        // diagnostic rather than attempting a conversion preserves both the
-        // invalid value and any user-owned formatting.
-        let Some(header_value) = value.as_str() else {
-            return ManagedHeaderInspection::Invalid;
-        };
-        if key.eq_ignore_ascii_case(CODEX_IMAGE_EXTENSION_HEADER) {
-            matching_keys.push((key.to_owned(), header_value == CODEX_IMAGE_EXTENSION_VALUE));
-        }
-    }
-
-    match matching_keys.as_slice() {
-        [] => ManagedHeaderInspection::Missing,
-        [(key, true)] => ManagedHeaderInspection::Controlled { key: key.clone() },
-        [(key, false)] => ManagedHeaderInspection::Conflict { key: key.clone() },
-        // Multiple case variants must never be normalized automatically, even
-        // when their values currently happen to agree.
-        [(key, _), ..] => ManagedHeaderInspection::Conflict { key: key.clone() },
-    }
-}
-
-fn websocket_state(provider_table: &dyn TableLike) -> (bool, Option<CodexConfigDiagnostic>) {
-    let Some(item) = provider_table.get("supports_websockets") else {
-        return (false, None);
-    };
-    match item.as_bool() {
-        Some(enabled) => (enabled, None),
-        None => (
-            false,
-            Some(codex_feature_diagnostic(
-                CODEX_FEATURE_INVALID_WEBSOCKET,
-                "supportsWebsockets",
-            )),
-        ),
-    }
-}
-
-fn image_extension_marker_is_complete(provider: &Provider) -> bool {
-    provider
-        .meta
-        .as_ref()
-        .and_then(|meta| meta.image_extension_configured)
-        == Some(true)
-}
-
-fn normalize_unfinished_image_extension_marker(provider: &mut Provider) {
-    if let Some(meta) = provider.meta.as_mut() {
-        if meta.image_extension_configured == Some(false) {
-            meta.image_extension_configured = None;
-        }
-    }
-}
-
-fn analyze_codex_provider_features_from_document(
-    provider: &Provider,
-    doc: &DocumentMut,
-    is_new: bool,
-) -> CodexProviderFeatureState {
-    let is_official = is_fixed_official_codex_provider(provider);
-    let image_extension_configured = image_extension_marker_is_complete(provider);
-    let missing_image_state = || {
-        if is_official {
-            CodexImageExtensionState::Off
-        } else if is_new && !image_extension_configured {
-            CodexImageExtensionState::On
-        } else if !image_extension_configured {
-            CodexImageExtensionState::LegacyPendingOn
-        } else {
-            CodexImageExtensionState::Off
-        }
-    };
-
-    let Some((_provider_id, provider_table)) = active_codex_provider_table(doc) else {
-        return CodexProviderFeatureState {
-            applicable: true,
-            image_extension: missing_image_state(),
-            websockets: CodexWebsocketFeatureState {
-                enabled: false,
-                compatible: true,
-                reason: None,
-            },
-            provider_table_found: false,
-            diagnostics: Vec::new(),
-        };
-    };
-
-    let header = inspect_managed_image_header(provider_table);
-    let (websocket_enabled, websocket_diagnostic) = websocket_state(provider_table);
-    let mut diagnostics = Vec::new();
-    if matches!(header, ManagedHeaderInspection::Invalid) {
-        diagnostics.push(codex_feature_diagnostic(
-            CODEX_FEATURE_INVALID_HEADER,
-            "httpHeaders",
-        ));
-    }
-    if let Some(diagnostic) = websocket_diagnostic {
-        diagnostics.push(diagnostic);
-    }
-
-    let image_extension = match header {
-        ManagedHeaderInspection::Missing => missing_image_state(),
-        ManagedHeaderInspection::Controlled { .. } => CodexImageExtensionState::On,
-        ManagedHeaderInspection::Conflict { key } => CodexImageExtensionState::Conflict { key },
-        ManagedHeaderInspection::Invalid => CodexImageExtensionState::Invalid {
-            code: CODEX_FEATURE_INVALID_HEADER.to_owned(),
-        },
-    };
-
-    CodexProviderFeatureState {
-        applicable: true,
-        image_extension,
-        websockets: CodexWebsocketFeatureState {
-            enabled: websocket_enabled,
-            compatible: true,
-            reason: None,
-        },
-        provider_table_found: true,
-        diagnostics,
-    }
-}
-
-/// Derive the Codex native-capability UI state from a provider draft. This is
-/// intentionally read-only and does not persist a migration marker or touch a
-/// user configuration file.
-pub fn analyze_codex_provider_features(
-    provider: &Provider,
-    is_new: bool,
-) -> CodexProviderFeatureState {
-    let config = codex_provider_config_text(provider);
-    match config.parse::<DocumentMut>() {
-        Ok(doc) => analyze_codex_provider_features_from_document(provider, &doc, is_new),
-        Err(_) => CodexProviderFeatureState {
-            applicable: false,
-            image_extension: CodexImageExtensionState::Invalid {
-                code: CODEX_FEATURE_INVALID_TOML.to_owned(),
-            },
-            websockets: CodexWebsocketFeatureState {
-                enabled: false,
-                compatible: false,
-                reason: None,
-            },
-            provider_table_found: false,
-            diagnostics: vec![codex_feature_diagnostic(
-                CODEX_FEATURE_INVALID_TOML,
-                "config",
-            )],
-        },
-    }
-}
-
-fn set_managed_image_header(provider_table: &mut dyn TableLike, enabled: bool) {
-    // An invalid header field cannot be preserved while applying an explicit
-    // image-toggle intent. The user chose this destructive repair boundary:
-    // enabling replaces it with the managed map, disabling removes it.
-    if matches!(
-        inspect_managed_image_header(provider_table),
-        ManagedHeaderInspection::Invalid
-    ) {
-        provider_table.remove("http_headers");
-    }
-
-    if provider_table.get("http_headers").is_none() {
-        if enabled {
-            let mut headers = toml_edit::InlineTable::new();
-            headers.insert(
-                CODEX_IMAGE_EXTENSION_HEADER,
-                CODEX_IMAGE_EXTENSION_VALUE.into(),
-            );
-            provider_table.insert(
-                "http_headers",
-                Item::Value(toml_edit::Value::InlineTable(headers)),
-            );
-        }
-        return;
-    }
-
-    let headers_empty = {
-        let Some(headers) = provider_table
-            .get_mut("http_headers")
-            .and_then(Item::as_table_like_mut)
-        else {
-            // The invalid case was removed above. Keep this defensive branch
-            // no-write instead of replacing an unexpected toml_edit shape.
-            return;
-        };
-        let matching_keys = headers
-            .iter()
-            .filter(|(key, _)| key.eq_ignore_ascii_case(CODEX_IMAGE_EXTENSION_HEADER))
-            .map(|(key, _)| key.to_owned())
-            .collect::<Vec<_>>();
-        for key in matching_keys {
-            headers.remove(&key);
-        }
-        if enabled {
-            headers.insert(
-                CODEX_IMAGE_EXTENSION_HEADER,
-                toml_edit::value(CODEX_IMAGE_EXTENSION_VALUE),
-            );
-        }
-        headers.is_empty()
-    };
-
-    if headers_empty {
-        provider_table.remove("http_headers");
-    }
-}
-
-fn set_provider_config_text(provider: &mut Provider, config_text: String) -> Result<(), AppError> {
-    let settings = provider.settings_config.as_object_mut().ok_or_else(|| {
-        AppError::localized(
-            "provider.codex.settings.not_object",
-            "Codex 配置必须是 JSON 对象",
-            "Codex configuration must be a JSON object",
-        )
-    })?;
-    settings.insert("config".to_owned(), Value::String(config_text));
-    Ok(())
-}
-
-fn feature_state_error(state: &CodexProviderFeatureState) -> Option<AppError> {
-    if state
-        .diagnostics
-        .iter()
-        .any(|diagnostic| diagnostic.code == CODEX_FEATURE_INVALID_TOML)
-    {
-        return Some(codex_feature_error(
-            CODEX_FEATURE_INVALID_TOML,
-            "Codex TOML 配置无效，无法保存原生能力设置",
-            "The Codex TOML configuration is invalid and native capabilities cannot be saved",
-        ));
-    }
-    None
-}
-
-/// Whole-document TOML failure is the only native-capability patch blocker.
-/// Field-level diagnostics remain visible but are repaired only when the user
-/// explicitly operates the corresponding switch.
-fn feature_state_error_for_patch(
-    state: &CodexProviderFeatureState,
-    _intent: &CodexProviderFeatureIntent,
-) -> Option<AppError> {
-    feature_state_error(state)
-}
-
-fn ensure_codex_feature_provider_table(
-    doc: &mut DocumentMut,
-    provider: &Provider,
-) -> Result<(String, bool), AppError> {
-    if let Some((provider_id, _)) = active_codex_provider_table(doc) {
-        return Ok((provider_id, false));
-    }
-
-    let is_official = is_fixed_official_codex_provider(provider);
-    let provider_id = if is_official {
-        FYAGENT_CODEX_MODEL_PROVIDER_ID.to_owned()
-    } else {
-        active_codex_model_provider_id(doc)
-            .unwrap_or_else(|| FYAGENT_CODEX_MODEL_PROVIDER_ID.to_owned())
-    };
-
-    if doc.get("model_providers").is_none() {
-        let mut providers = toml_edit::Table::new();
-        providers.set_implicit(true);
-        doc["model_providers"] = Item::Table(providers);
-    }
-    let providers = doc
-        .get_mut("model_providers")
-        .and_then(Item::as_table_like_mut)
-        .ok_or_else(|| {
-            codex_feature_error(
-                CODEX_FEATURE_INVALID_TOML,
-                "Codex model_providers 必须是可编辑的表",
-                "Codex model_providers must be an editable table",
-            )
-        })?;
-
-    let created_provider_table = providers.get(&provider_id).is_none();
-    if created_provider_table {
-        let mut table = toml_edit::Table::new();
-        table["name"] = toml_edit::value(if is_official {
-            "OpenAI"
-        } else {
-            provider.name.trim()
-        });
-        if is_official {
-            table["requires_openai_auth"] = toml_edit::value(true);
-        }
-        table["wire_api"] = toml_edit::value("responses");
-        providers.insert(&provider_id, Item::Table(table));
-    }
-
-    doc["model_provider"] = toml_edit::value(&provider_id);
-    Ok((provider_id, is_official && created_provider_table))
-}
-
-fn generated_official_provider_table_is_safe_to_remove(doc: &DocumentMut) -> bool {
-    if active_codex_model_provider_id(doc).as_deref() != Some(FYAGENT_CODEX_MODEL_PROVIDER_ID) {
-        return false;
-    }
-    let Some((_, table)) = active_codex_provider_table(doc) else {
-        return false;
-    };
-    for (key, _) in table.iter() {
-        if !matches!(key, "name" | "requires_openai_auth" | "wire_api") {
-            return false;
-        }
-    }
-    table.get("name").and_then(Item::as_str) == Some("OpenAI")
-        && table.get("requires_openai_auth").and_then(Item::as_bool) == Some(true)
-        && table.get("wire_api").and_then(Item::as_str) == Some("responses")
-}
-
-fn remove_generated_official_provider_table(doc: &mut DocumentMut) -> bool {
-    if !generated_official_provider_table_is_safe_to_remove(doc) {
-        return false;
-    }
-
-    doc.as_table_mut().remove("model_provider");
-    let providers_empty = doc
-        .get_mut("model_providers")
-        .and_then(Item::as_table_like_mut)
-        .is_some_and(|providers| {
-            providers.remove(FYAGENT_CODEX_MODEL_PROVIDER_ID);
-            providers.is_empty()
-        });
-    if providers_empty {
-        doc.as_table_mut().remove("model_providers");
-    }
-    true
-}
-
-/// Patch a form-only Codex provider TOML draft without rewriting unrelated
-/// formatting. The returned TOML must be persisted by the caller through the
-/// normal provider save path; this function never writes a live config file.
-pub fn patch_codex_provider_features(
-    provider: &Provider,
-    intent: &CodexProviderFeatureIntent,
-    is_new: bool,
-) -> Result<CodexProviderFeaturePatchResult, AppError> {
-    let config = codex_provider_config_text(provider);
-    let mut doc = config.parse::<DocumentMut>().map_err(|_| {
-        codex_feature_error(
-            CODEX_FEATURE_INVALID_TOML,
-            "Codex TOML 配置无效，无法修改原生能力设置",
-            "The Codex TOML configuration is invalid and native capabilities cannot be changed",
-        )
-    })?;
-    let state = analyze_codex_provider_features_from_document(provider, &doc, is_new);
-    if let Some(error) = feature_state_error_for_patch(&state, intent) {
-        return Err(error);
-    }
-
-    let needs_provider_table = intent.image_extension == Some(true)
-        || intent.websockets == Some(true)
-        || active_codex_provider_table(&doc).is_some();
-    let mut generated_provider_marker = None;
-    let provider_id = if needs_provider_table {
-        let (provider_id, generated_official) =
-            ensure_codex_feature_provider_table(&mut doc, provider)?;
-        if generated_official {
-            generated_provider_marker = Some(true);
-        }
-        Some(provider_id)
-    } else {
-        None
-    };
-
-    let is_official = is_fixed_official_codex_provider(provider);
-    let image_extension_configured =
-        (intent.image_extension.is_some() && !is_official).then_some(true);
-    if let Some(provider_id) = provider_id.as_deref() {
-        let provider_table =
-            active_codex_provider_table_mut(&mut doc, provider_id).ok_or_else(|| {
-                codex_feature_error(
-                    CODEX_FEATURE_INVALID_TOML,
-                    "Codex Provider 表不可编辑",
-                    "The Codex provider table is not editable",
-                )
-            })?;
-        if let Some(enabled) = intent.image_extension {
-            set_managed_image_header(provider_table, enabled);
-        }
-        if let Some(enabled) = intent.websockets {
-            if enabled {
-                provider_table.insert("supports_websockets", toml_edit::value(true));
-            } else {
-                provider_table.remove("supports_websockets");
-            }
-        }
-    }
-
-    let owns_generated_official_table = is_official
-        && (generated_official_provider_marker(provider)
-            || generated_provider_marker == Some(true));
-    if owns_generated_official_table && remove_generated_official_provider_table(&mut doc) {
-        generated_provider_marker = Some(false);
-    }
-
-    let toml_text = doc.to_string();
-    // Reparse before reporting a successful patch. This catches an unexpected
-    // toml_edit construction defect without serializing or logging the draft.
-    let parsed = toml_text.parse::<DocumentMut>().map_err(|_| {
-        codex_feature_error(
-            CODEX_FEATURE_INVALID_TOML,
-            "Codex TOML 配置无效，无法修改原生能力设置",
-            "The Codex TOML configuration is invalid and native capabilities cannot be changed",
-        )
-    })?;
-    let mut patched_provider = provider.clone();
-    set_provider_config_text(&mut patched_provider, toml_text.clone())?;
-    if image_extension_configured.is_some() {
-        // This is a form-only clone. It lets the returned state describe the
-        // explicit user action without persisting the marker before save.
-        patched_provider
-            .meta
-            .get_or_insert_with(ProviderMeta::default)
-            .image_extension_configured = Some(true);
-    }
-    if let Some(generated) = generated_provider_marker {
-        patched_provider
-            .meta
-            .get_or_insert_with(ProviderMeta::default)
-            .codex_native_capabilities_generated_provider = generated.then_some(true);
-    }
-    Ok(CodexProviderFeaturePatchResult {
-        state: analyze_codex_provider_features_from_document(&patched_provider, &parsed, is_new),
-        toml_text,
-        image_extension_configured,
-        codex_native_capabilities_generated_provider: generated_provider_marker,
-    })
-}
-
-/// Enforce only the whole-document editability invariant on Codex provider
-/// write paths. Field-level native-capability diagnostics are deliberately
-/// non-blocking until the user operates the corresponding switch.
-pub fn validate_codex_provider_features(provider: &Provider) -> Result<(), AppError> {
-    let state = analyze_codex_provider_features(provider, false);
-    if let Some(error) = feature_state_error(&state) {
-        return Err(error);
-    }
-    Ok(())
-}
-
-/// Apply the small, deferred image default at the moment a non-official
-/// provider is actually saved. Historical records are never rewritten merely
-/// because they are listed or opened. Conflicting or invalid header fields are
-/// preserved by an unrelated save and repaired only by an explicit toggle.
-pub fn prepare_codex_provider_features_for_save(
-    provider: &mut Provider,
-    is_new: bool,
-) -> Result<(), AppError> {
-    // `false` was never a completed choice. Normalize manually constructed
-    // drafts as well as historical records before all feature decisions, so
-    // a successful save cannot preserve the stale completion marker.
-    normalize_unfinished_image_extension_marker(provider);
-    if let Some(meta) = provider.meta.as_mut() {
-        if meta.codex_native_capabilities_generated_provider == Some(false) {
-            meta.codex_native_capabilities_generated_provider = None;
-        }
-    }
-    validate_codex_provider_features(provider)?;
-    let state = analyze_codex_provider_features(provider, is_new);
-    if !state.applicable {
-        return Ok(());
-    }
-    if matches!(
-        state.image_extension,
-        CodexImageExtensionState::Conflict { .. } | CodexImageExtensionState::Invalid { .. }
-    ) {
-        return Ok(());
-    }
-
-    // A newly created eligible provider starts with the feature logically on,
-    // but the renderer records an explicit first-save choice before it calls
-    // this function.  Do not let the default overwrite an explicit off choice
-    // simply because the provider has not been persisted yet.
-    let has_image_extension_marker = image_extension_marker_is_complete(provider);
-    let should_apply_default = !is_fixed_official_codex_provider(provider)
-        && ((is_new && !has_image_extension_marker)
-            || matches!(
-                state.image_extension,
-                CodexImageExtensionState::LegacyPendingOn
-            ));
-    if should_apply_default {
-        let result = patch_codex_provider_features(
-            provider,
-            &CodexProviderFeatureIntent {
-                image_extension: Some(true),
-                websockets: None,
-            },
-            false,
-        )?;
-        set_provider_config_text(provider, result.toml_text)?;
-        provider
-            .meta
-            .get_or_insert_with(ProviderMeta::default)
-            .image_extension_configured = Some(true);
-        return Ok(());
-    }
-
-    if !is_fixed_official_codex_provider(provider)
-        && matches!(state.image_extension, CodexImageExtensionState::On)
-        && provider
-            .meta
-            .as_ref()
-            .and_then(|meta| meta.image_extension_configured)
-            != Some(true)
-    {
-        provider
-            .meta
-            .get_or_insert_with(ProviderMeta::default)
-            .image_extension_configured = Some(true);
-    }
-    Ok(())
-}
-
-fn codex_provider_websocket_enabled(provider: &Provider) -> bool {
-    codex_provider_config_text(provider)
-        .parse::<DocumentMut>()
-        .ok()
-        .and_then(|doc| {
-            active_codex_provider_table(&doc)
-                .and_then(|(_, table)| table.get("supports_websockets").and_then(Item::as_bool))
-        })
-        == Some(true)
-}
-
-fn is_gpt_model_id(model: &str) -> Option<bool> {
-    let model = model.trim();
-    if model.is_empty() {
-        return None;
-    }
-    let basename = model.rsplit('/').next().unwrap_or_default().trim();
-    Some(basename.to_ascii_lowercase().starts_with("gpt-"))
-}
-
-/// Derive non-sensitive warnings for a successful Codex provider add/update.
-/// The caller supplies the current takeover state so this pure classifier is
-/// shared by commands and directly testable without touching user files.
-pub fn codex_provider_save_warning_codes(
-    provider: &Provider,
-    proxy_takeover_active: bool,
-) -> Vec<String> {
-    if !codex_provider_websocket_enabled(provider) {
-        return Vec::new();
-    }
-
-    let mut has_non_gpt_model = false;
-    if let Ok(doc) = codex_provider_config_text(provider).parse::<DocumentMut>() {
-        has_non_gpt_model = ["model", "review_model"].into_iter().any(|field| {
-            doc.get(field)
-                .and_then(Item::as_str)
-                .and_then(is_gpt_model_id)
-                == Some(false)
-        });
-    }
-    if !has_non_gpt_model {
-        has_non_gpt_model = provider
-            .settings_config
-            .get("modelCatalog")
-            .and_then(|catalog| catalog.get("models"))
-            .and_then(Value::as_array)
-            .is_some_and(|models| {
-                models.iter().any(|entry| {
-                    entry
-                        .get("model")
-                        .and_then(Value::as_str)
-                        .and_then(is_gpt_model_id)
-                        == Some(false)
-                })
-            });
-    }
-
-    let mut warnings = Vec::new();
-    if has_non_gpt_model {
-        warnings.push(CODEX_WEBSOCKET_NON_GPT_MODEL_WARNING.to_owned());
-    }
-    if proxy_takeover_active {
-        warnings.push(CODEX_WEBSOCKET_PROXY_MAY_BE_UNSUPPORTED_WARNING.to_owned());
-    }
-    warnings
 }
 
 /// Write only Codex `config.toml` for provider switching.
@@ -1852,222 +1098,6 @@ pub fn prepare_codex_config_text_with_model_catalog(
         let disable_web_search = profile == CodexCatalogToolProfile::Anthropic;
         set_codex_native_web_search_field(&config_text, disable_web_search)
     }
-}
-
-/// Reverse of `prepare_codex_config_text_with_model_catalog`: read the
-/// fyagent–maintained catalog file referenced by `~/.codex/config.toml` and
-/// convert it back into the simplified shape the frontend table uses:
-/// `{ "models": [{ "model", "displayName"?, "contextWindow"?, hidden overrides... }, ...] }`.
-///
-/// We only reverse-parse catalogs whose `model_catalog_json` path is the
-/// fyagent–generated file (identified by filename
-/// `fyagent-model-catalog.json`). A user-managed external catalog file is
-/// left alone — surfacing its richer structure as the simplified table would
-/// be a downgrade we can't safely round-trip.
-///
-/// `displayName`, `contextWindow`, and `inputModalities` are omitted from the
-/// returned entry when the on-disk value matches the fallback that
-/// `codex_model_catalog_from_settings` injects for unset inputs (slug for
-/// display_name, `model_context_window` or 128_000 for context_window, and the
-/// shared confirmed-text-only inference for input modalities). This preserves
-/// the "user left it blank" intent across round-trip; an unavoidable edge case
-/// is that a user-typed value that happens to equal the fallback also collapses
-/// to blank, but the next save writes the same fallback so behavior is stable.
-///
-/// All failure modes (missing file, parse error, no `model_catalog_json`,
-/// entries without `slug`) collapse to `Ok(None)` so callers can treat this
-/// as best-effort enrichment without making `read_live_settings` brittle.
-/// 模型目录文件读取上限（32 MiB）。目录 JSON 正常只有几百 KiB；超过则视为异常，
-/// 避免指向外部大文件时耗尽内存。
-const MAX_CODEX_CATALOG_BYTES: u64 = 32 * 1024 * 1024;
-
-pub fn read_codex_model_catalog_simplified_from_live() -> Result<Option<Value>, AppError> {
-    let config_text = read_codex_config_text()?;
-    let config_dir = get_codex_config_dir();
-    let Some(catalog_path) = resolve_fyagent_catalog_path(&config_text, &config_dir) else {
-        return Ok(None);
-    };
-    if !catalog_path.exists() {
-        return Ok(None);
-    }
-    let catalog_text = match read_limited_string(&catalog_path, MAX_CODEX_CATALOG_BYTES) {
-        Ok(text) => text,
-        Err(error) => {
-            log::warn!(
-                "拒绝读取越界或过大的 Codex 模型目录 {}: {error}",
-                catalog_path.display()
-            );
-            return Ok(None);
-        }
-    };
-    Ok(build_simplified_catalog_from_texts(
-        &config_text,
-        &catalog_text,
-    ))
-}
-
-/// 安全地读取文件为字符串，并在超过字节上限时返回错误。
-pub(crate) fn read_limited_string(path: &Path, max_bytes: u64) -> Result<String, AppError> {
-    let metadata = fs::metadata(path).map_err(|error| AppError::io(path, error))?;
-    if metadata.len() > max_bytes {
-        return Err(AppError::Config(format!(
-            "文件 {} 超过大小上限 {} 字节",
-            path.display(),
-            max_bytes
-        )));
-    }
-    fs::read_to_string(path).map_err(|error| AppError::io(path, error))
-}
-
-/// Read the FyAgent Codex model catalog file with a size cap.
-pub(crate) fn read_codex_model_catalog_text(path: &Path) -> Result<String, AppError> {
-    read_limited_string(path, MAX_CODEX_CATALOG_BYTES)
-}
-
-/// Given `config.toml` text, resolve the on-disk path of the FyAgent-owned
-/// catalog file (returns `None` if `model_catalog_json` is absent or points at
-/// a file we don't own). Relative paths are resolved under `base_dir`;
-/// absolute paths must still be inside `base_dir`.
-pub(crate) fn resolve_fyagent_catalog_path(config_text: &str, base_dir: &Path) -> Option<PathBuf> {
-    if config_text.trim().is_empty() {
-        return None;
-    }
-    let doc = config_text.parse::<DocumentMut>().ok()?;
-    let catalog_path_str = doc
-        .get("model_catalog_json")
-        .and_then(|item| item.as_str())
-        .map(str::trim)
-        .filter(|s| !s.is_empty())?;
-
-    let referenced_path = Path::new(catalog_path_str);
-    let is_fyagent_owned = referenced_path.file_name().and_then(|name| name.to_str())
-        == Some(FYAGENT_CODEX_MODEL_CATALOG_FILENAME);
-    if !is_fyagent_owned {
-        return None;
-    }
-
-    // Windows 的 `Path` 不会把前导 `/` 识别成绝对路径。这里仍显式按绝对路径
-    // 处理这类跨平台配置值，让下方包含性校验 fail closed，而不是把目录外路径
-    // 误拼到配置目录下。下一次切换供应商时，写入侧会重新落一个裸文件名。
-    let is_unix_absolute = catalog_path_str.starts_with('/');
-    let resolved = if referenced_path.is_absolute() || is_unix_absolute {
-        referenced_path.to_path_buf()
-    } else {
-        base_dir.join(referenced_path)
-    };
-
-    if !path_is_within(base_dir, &resolved) {
-        log::warn!(
-            "Codex model_catalog_json 指向配置目录外: {}（允许目录: {}）",
-            resolved.display(),
-            base_dir.display()
-        );
-        return None;
-    }
-
-    // 词法包含不等于运行时包含：配置目录内的符号链接（如 ~/.codex/link ->
-    // /etc）能让 `link/fyagent-model-catalog.json` 通过上面的检查，读取却
-    // 落到目录外。文件存在时把真实路径 canonicalize 出来再校验一次，并把
-    // canonical 路径返回给调用方——后续读取不再经过 symlink 组件。
-    if resolved.exists() {
-        let canonical = match fs::canonicalize(&resolved) {
-            Ok(path) => path,
-            Err(error) => {
-                log::warn!(
-                    "Codex model_catalog_json canonicalize 失败: {}: {error}",
-                    resolved.display()
-                );
-                return None;
-            }
-        };
-        // base 同样 canonicalize，保证两侧前缀一致（Windows \\?\、
-        // macOS /tmp -> /private/tmp）；base 失败时退回词法 base——
-        // 词法 base 与 canonical 路径比较只会误拒（退化为不读），不会误放。
-        let canonical_base = fs::canonicalize(base_dir).unwrap_or_else(|_| base_dir.to_path_buf());
-        if !path_is_within(&canonical_base, &canonical) {
-            log::warn!(
-                "Codex model_catalog_json 经符号链接解析到配置目录外: {} -> {}（允许目录: {}）",
-                resolved.display(),
-                canonical.display(),
-                canonical_base.display()
-            );
-            return None;
-        }
-        return Some(canonical);
-    }
-
-    Some(resolved)
-}
-
-/// Pure reverse-parsing core: convert Codex catalog JSON text back into the
-/// frontend's simplified model-mapping shape. Returns `None` when the catalog
-/// is unparseable, has no `models` array, or yields zero valid entries.
-fn build_simplified_catalog_from_texts(config_text: &str, catalog_text: &str) -> Option<Value> {
-    let catalog: Value = serde_json::from_str(catalog_text).ok()?;
-    let models = catalog.get("models").and_then(|m| m.as_array())?;
-
-    let default_context_window =
-        extract_codex_top_level_u64(config_text, "model_context_window").unwrap_or(128_000);
-
-    let mut entries = Vec::with_capacity(models.len());
-    for entry in models {
-        let Some(model) = entry
-            .get("slug")
-            .and_then(|v| v.as_str())
-            .map(str::trim)
-            .filter(|s| !s.is_empty())
-        else {
-            continue;
-        };
-
-        let mut obj = serde_json::Map::new();
-        obj.insert("model".to_string(), json!(model));
-
-        if let Some(display_name) = entry
-            .get("display_name")
-            .and_then(|v| v.as_str())
-            .map(str::trim)
-            .filter(|s| !s.is_empty() && *s != model)
-        {
-            obj.insert("displayName".to_string(), json!(display_name));
-        }
-
-        if let Some(context_window) = entry
-            .get("context_window")
-            .and_then(|v| v.as_u64())
-            .filter(|v| *v > 0 && *v != default_context_window)
-        {
-            obj.insert("contextWindow".to_string(), json!(context_window));
-        }
-
-        // Preserve native-profile per-row overrides so a DB-SSOT-missing
-        // fallback round-trip doesn't silently drop them.
-        if let Some(parallel) = entry
-            .get("supports_parallel_tool_calls")
-            .and_then(|v| v.as_bool())
-        {
-            obj.insert("supportsParallelToolCalls".to_string(), json!(parallel));
-        }
-        if let Some(modalities) = entry.get("input_modalities").and_then(|v| v.as_array()) {
-            let mods: Vec<String> = modalities
-                .iter()
-                .filter_map(|m| m.as_str())
-                .map(str::to_string)
-                .collect();
-            let inferred = codex_catalog_input_modalities(model, None);
-            if !mods.is_empty() && mods != inferred {
-                obj.insert("inputModalities".to_string(), json!(mods));
-            }
-        }
-
-        entries.push(Value::Object(obj));
-    }
-
-    if entries.is_empty() {
-        return None;
-    }
-
-    Some(json!({ "models": entries }))
 }
 
 /// Decide the `config.toml` text to write during a takeover-off restore,
